backend/requirements.txt
# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6
pydantic==2.6.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap==3.4.3

# Database
SQLAlchemy==2.0.27
alembic==1.13.1
aiosqlite==0.19.0

# Authentication
python-ldap==3.4.3
ldap3==2.9.1

# Validation
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.0
httpx==0.24.1

pydantic-settings==2.0.2
